Support >30 fps
The VPU Middleware doesn't allow settings for >30fps, but if we tell
it that it's 30fps and adjust the bitrate accordingly it works ok.
The timestamps in the ES are of course wrong, but most applications
ignore these anyway as they are generally unreliable. Since the
decoder API doesn;t provide the timestamps, there shouldn't be any
porblems.

Note: Set the 'frame number resolution' property to the fps.